Flytter UtpekingVurderingSelector til flytSelectors for å unngå sirkulære avhengigheter
