+++ src/ducks/flyt/selectors.js
@@ -1,9 +1,12 @@
 import { createSelector } from 'reselect';
 
+import MKV from '../../melosyskodeverk';
 import * as KV from '../../kodeverk';
 
 import { avklartefaktaSelectors } from '../avklartefakta';
 import { vilkarSelectors } from '../vilkar';
+import { behandlingerSelectors } from '../behandlinger';
+import { behandlingsresultatSelectors } from '../behandlingsresultat';
 
 export const ErIArtikkel11_4Selector = createSelector(
   state => vilkarSelectors.art11_4_1(state),
@@ -24,4 +27,15 @@ export const HarOffentligTjenesteAnnetLandSelector = createSelector(
 export const HarLonnetArbeidAnnetLand = createSelector(
   state => avklartefaktaSelectors.LoennetArbeidAntallLandFaktaVerdiSelector(state),
   loennetArbeidAntallLand => loennetArbeidAntallLand === KV.Koder.LoennetArbeidAntallLand.ETT_ANNET_LAND
+);
+
+export const UtpekingVurderingSelector = createSelector(
+  state => behandlingerSelectors.BehandlingstemaKodeSelector(state),
+  state => behandlingsresultatSelectors.UtfallRegistreringUnntakSelector(state),
+  state => behandlingsresultatSelectors.UtfallUtpekingSelector(state),
+  (behandlingstema, utfallRegistreringUnntak, utfallUtpeking) => (
+    behandlingstema === MKV.Koder.behandlinger.behandlingstema.BESLUTNING_LOVVALG_ANNET_LAND
+      ? utfallRegistreringUnntak
+      : utfallUtpeking
+  )
 );
